- rework product deletion flow

+++ ProductsService/src/main/java/com/gostavdev/commercify/productsservice/services/ProductService.java
@@ -5,6 +5,7 @@ import com.gostavdev.commercify.productsservice.dto.ProductDTOMapper;
 import com.gostavdev.commercify.productsservice.entities.ProductEntity;
 import com.gostavdev.commercify.productsservice.repositories.ProductRepository;
 import com.gostavdev.commercify.productsservice.requests.ProductRequest;
+import com.stripe.Stripe;
 import com.stripe.exception.StripeException;
 import com.stripe.model.Product;
 import com.stripe.param.ProductCreateParams;
@@ -43,25 +44,27 @@ public class ProductService {
                 .active(true)
                 .build();
 
-        try {
-            long amountInCents = (long) (productEntity.getUnitPrice() * 100);
+        if (!Stripe.apiKey.equals("sk_missing")) {
+            try {
+                long amountInCents = (long) (productEntity.getUnitPrice() * 100);
 
-            ProductCreateParams.DefaultPriceData defaultPriceData = ProductCreateParams.DefaultPriceData.builder()
-                    .setCurrency(productEntity.getCurrency())
-                    .setUnitAmount(amountInCents).build();
+                ProductCreateParams.DefaultPriceData defaultPriceData = ProductCreateParams.DefaultPriceData.builder()
+                        .setCurrency(productEntity.getCurrency())
+                        .setUnitAmount(amountInCents).build();
 
-            ProductCreateParams params =
-                    ProductCreateParams.builder()
-                            .setName(productEntity.getName())
-                            .setDescription(productEntity.getDescription())
-                            .setDefaultPriceData(defaultPriceData)
-                            .build();
-            Product stripeProduct = Product.create(params);
+                ProductCreateParams params =
+                        ProductCreateParams.builder()
+                                .setName(productEntity.getName())
+                                .setDescription(productEntity.getDescription())
+                                .setDefaultPriceData(defaultPriceData)
+                                .build();
+                Product stripeProduct = Product.create(params);
 
-            productEntity.setStripeId(stripeProduct.getId());
+                productEntity.setStripeId(stripeProduct.getId());
 
-        } catch (StripeException e) {
-            throw new RuntimeException("Stripe error", e);
+            } catch (StripeException e) {
+                throw new RuntimeException("Stripe error", e);
+            }
         }
 
         ProductEntity savedProduct = productRepository.save(productEntity);
@@ -70,37 +73,53 @@ public class ProductService {
     }
 
     public boolean deleteProduct(Long id, boolean forceDeletion) throws RuntimeException {
+        System.out.println("Stripe key: " + Stripe.apiKey);
         ProductEntity productEnt = productRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Product not found"));
 
-        Product resource;
-        try {
-            resource = Product.retrieve(productEnt.getStripeId());
-        } catch (StripeException e) {
-            productRepository.deleteById(id);
-            return true;
+        if (!productEnt.getActive()) {
+            throw new RuntimeException("Product is already deleted");
         }
 
-        if (forceDeletion) {
-            productRepository.deleteById(id);
-        } else {
+        if (!Stripe.apiKey.equals("sk_missing")) {
+            return deleteProductFromStripe(productEnt, forceDeletion);
+        }
+
+        if (Stripe.apiKey.equals("sk_missing") && productEnt.getStripeId() != null) {
+            throw new RuntimeException("Cant delete product from stripe without stripe key");
+        }
+
+        if (!forceDeletion) {
+            System.out.println("Deactivating product");
             productEnt.setActive(false);
             productRepository.save(productEnt);
-        }
-
-        if (!resource.getActive()) {
-            throw new RuntimeException("Product has already been deleted");
-        }
-
-        ProductUpdateParams params =
-                ProductUpdateParams.builder().setActive(false).build();
-
-        try {
-            resource.update(params);
-        } catch (StripeException e) {
             return false;
         }
 
+        productRepository.deleteById(id);
+        System.out.println("Product deleted");
         return true;
+    }
+
+    private boolean deleteProductFromStripe(ProductEntity productEnt, boolean forceDeletion) {
+        try {
+            if (forceDeletion) {
+                Product stripeProduct = Product.retrieve(productEnt.getStripeId());
+                stripeProduct.delete();
+                productRepository.deleteById(productEnt.getProductId());
+                return true;
+            }
+
+            ProductUpdateParams params = ProductUpdateParams.builder()
+                    .setActive(false)
+                    .build();
+            Product stripeProduct = Product.retrieve(productEnt.getStripeId());
+            stripeProduct.update(params);
+            productEnt.setActive(false);
+            productRepository.save(productEnt);
+            return false;
+        } catch (StripeException e) {
+            throw new RuntimeException("Stripe error", e);
+        }
     }
 
     public List<ProductDTO> saveProducts(List<ProductRequest> request) {
